Minimum zoom limit.

+++ src/assets/browser/src/controller/input/direct.rs
@@ -35,7 +35,7 @@ fn custom_movement_event(dir: &str, unit: &str, v: &JSONValue) -> Event {
 
 fn custom_zoom_event(kind: &str, v: &JSONValue) -> Event {
     if let JSONValue::Number(quant) = v {
-        let quant = quant.as_f64().unwrap() as f32;
+        let quant = quant.as_f64().unwrap();
         match kind {
             "by" => {
                 Event::Zoom(quant)
